Fix hooks exports
